Split outgoing and stored request representation into separate classes CIRC-342
